<기능> 재고, 발주 , 인사 검색 기능 추가 <오류> 발주현황 검색 오류 발생

+++ src/main/java/com/core/erp/dto/PurchaseOrderItemDTO.java
@@ -12,7 +12,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class PurchaseOrderItemDTO {
 
     private int itemId;
-    private Integer orderId; // FK (id만 관리)
+    private Long orderId; // FK (id만 관리)
     private Integer productId; // FK (id만 관리)
     private String productName;
     private int orderQuantity;
